% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_attr}
\alias{add_attr}
\title{Add attribute to abject}
\usage{
add_attr(var, attribute, name)
}
\arguments{
\item{var}{A tibble}

\item{attribute}{An object}

\item{name}{A character name of the attribute}
}
\value{
A tibble with an additional attribute
}
\description{
Add attribute to abject
}
\keyword{internal}
