% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_brush_data}
\alias{demo_brush_data}
\title{Demo brush data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 3 columns.
}
\usage{
demo_brush_data
}
\description{
Demo brush data
}
\keyword{data}
