% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_select_data}
\alias{demo_select_data}
\title{Demo select data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23 rows and 4 columns.
}
\usage{
demo_select_data
}
\description{
Demo select data
}
\keyword{data}
