% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_chr.numeric}
\alias{gate_chr.numeric}
\title{gate_chr}
\usage{
\method{gate_chr}{numeric}(
  .dim1,
  .dim2,
  .color = NULL,
  .shape = NULL,
  .size = NULL,
  opacity = 1,
  how_many_gates = 1,
  .group_by = NULL,
  gate_list = NULL,
  ...
)
}
\description{
gate_chr
}
