% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_custom}
\alias{gate_custom}
\title{Interactively gate data with a custom scatter plot}
\usage{
gate_custom(plot)
}
\arguments{
\item{plot}{A ggplot object. Must contain a row index in the `.key` column set as key.}
}
\value{
A vector of lists, recording the gates each X and Y coordinate pair is within. A record 
of the selected points is stored in `tidygate_env$select_data` and a record of the gates is 
stored in `tidygate_env$brush_data`.
}
\description{
Launch an interactive scatter plot, based on a user-defined `ggplot2`. Points on this plot can
then be gated using multiple lasso selections.
}
\examples{
\dontrun{
library(dplyr)
library(ggplot2)

scaled_plot <- 
  mtcars |> 
  mutate(.key = row_number()) |>
  ggplot(aes(x = mpg, y = wt, key = .key)) + 
  scale_y_log10() +
  geom_point() +
  theme_dark()
  
mtcars |>
  mutate(selected = gate_custom(custom_plot = scaled_plot)) |>
  print()
}
}
