% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_int}
\alias{gate_int}
\title{gate_int}
\usage{
gate_int(
  .dim1,
  .dim2,
  .color = NULL,
  .shape = NULL,
  .size = NULL,
  opacity = 1,
  how_many_gates = 1,
  .group_by = NULL,
  gate_list = NULL,
  ...
)
}
\description{
gate_int
}
