% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_interactive}
\alias{gate_interactive}
\title{Interactively gate data with a simple scatter plot}
\usage{
gate_interactive(x, y, colour = NULL, shape = NULL, alpha = 1, size = 2)
}
\arguments{
\item{x}{A vector representing the X dimension.}

\item{y}{A vector representing the Y dimension.}

\item{colour}{A vector representing the point colour. Or, a colour code string compatible
with ggplot2.}

\item{shape}{A vector representing the point shape, coercible to a factor. Or, a shape code 
numeric compatible with ggplot2 (0-127)}

\item{alpha}{A vector representing the point alpha, coercible to a factor. Or, an alpha numeric 
compatible with ggplot2 (0-1).}

\item{size}{A vector representing the point size, coercible to a factor. Or, a size numeric 
compatible with ggplot2 (0-20).}
}
\value{
A vector of lists, recording the gates each X and Y coordinate pair is within. A record 
of the selected points is stored in `tidygate_env$select_data` and a record of the gates is 
stored in `tidygate_env$brush_data`.
}
\description{
Launch an interactive scatter plot, based on the input parameters. Points on this plot can
then be gated using multiple lasso selections.
}
\examples{
\dontrun{
mtcars |>
  mutate(selected = gate_interactive(x = mpg, y = wt, shape = am))
}
}
