% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chr_int.R
\name{gate_interactive_chr_int}
\alias{gate_interactive_chr_int}
\title{Get points within a user drawn gate}
\usage{
gate_interactive_chr_int(
  .data,
  .dim1,
  .dim2,
  .color = NA,
  .shape = NULL,
  .size = NULL,
  opacity = 1,
  how_many_gates = 1,
  is_size_fixed,
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.dim1}{A column symbol. The x dimension}

\item{.dim2}{A column symbol. The y dimension}

\item{.color}{A column symbol. Color of points}

\item{.shape}{A column symbol. Shape of points}

\item{.size}{A column symbol. Size of points}

\item{opacity}{A number between 0 and 1. The opacity level of the data points}

\item{how_many_gates}{An integer. The number of gates to label}

\item{...}{Further parameters passed to the function gatepoints::fhs}

\item{gate_list}{A list of gates. It is returned by gate function as attribute \"gate\". If you want to create this list yourself, each element of the list is a data frame with x and y columns. Each row is a coordinate. The order matter.}
}
\value{
A tibble with additional columns
}
\description{
Get points within a user drawn gate
}
\keyword{internal}
