% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_chr_int.R
\name{gate_programmatic_chr_int}
\alias{gate_programmatic_chr_int}
\title{Get points within a user drawn gate}
\usage{
gate_programmatic_chr_int(.data, .dim1, .dim2, gate_list)
}
\arguments{
\item{.data}{A tibble}

\item{.dim1}{A column symbol. The x dimension}

\item{.dim2}{A column symbol. The y dimension}

\item{gate_list}{A list of gates. Each element of the list is a data frame with x and y columns. Each row is a coordinate. The order matter.}
}
\value{
A tibble with additional columns
}
\description{
Get points within a user drawn gate
}
\keyword{internal}
