% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_OLD.R
\name{gate_programmatic_old}
\alias{gate_programmatic_old}
\title{Get points within a user drawn gate}
\usage{
gate_programmatic_old(
  .data,
  .element,
  .dim1,
  .dim2,
  gate_list,
  name = "gate",
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.dim1}{A column symbol. The x dimension}

\item{.dim2}{A column symbol. The y dimension}

\item{gate_list}{A list of gates. Each element of the list is a data frame with x and y columns. Each row is a coordinate. The order matter.}

\item{name}{A character string. The name of the new column}

\item{...}{Further parameters passed to the function gatepoints::fhs}
}
\value{
A tibble with additional columns
}
\description{
Get points within a user drawn gate
}
\keyword{internal}
