% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_simple}
\alias{gate_simple}
\title{Interactively gate data with a simple scatter plot}
\usage{
gate_simple(
  dimension_x,
  dimension_y,
  colour = NULL,
  shape = NULL,
  alpha = 1,
  size = 1
)
}
\arguments{
\item{dimension_x}{A column symbol representing the X dimension.}

\item{dimension_y}{A column symbol representing the Y dimension.}

\item{colour}{A column symbol representing the point colour.}

\item{shape}{A column symbol representing the point shape. Must be a factor.}

\item{alpha}{A numeric value representing the opacity of points, with 1 being completely opaque and 0 being completely
transparent.}

\item{size}{A numeric value representing the size of points.}
}
\value{
A vector of lists, recording the gates each X and Y coordinate pair is within. A record 
of the selected points is stored in `tidygate_env$select_data` and a record of the gates is 
stored in `tidygate_env$brush_data`.
}
\description{
Launch an interactive scatter plot, based on the input parameters. Points on this plot can
then be gated using multiple lasso selections.
}
\examples{
\dontrun{
library(dplyr)

mtcars |>
  dplyr::mutate(selected = gate_simple(dimension_x = mpg, dimension_y = wt, alpha = 0.5)) |>
  print()
}
}
