% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_app.R
\name{server}
\alias{server}
\title{Run Shiny App for interactive gating}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Server input parameter}

\item{output}{Server output parameter}

\item{session}{Server session parameter}
}
\value{
NA
}
\description{
Run Shiny App for interactive gating
}
