% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidygate_data}
\alias{tidygate_data}
\title{Example data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2240 rows and 8 columns.
}
\usage{
tidygate_data
}
\description{
Example data set
}
\keyword{internal}
