% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gate_app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Create Shiny App UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
Fluid UI container
}
\description{
Create Shiny App UI
}
\keyword{datasets}
