% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_interactive}
\alias{gate_interactive}
\title{Interactively gate data}
\usage{
gate_interactive(dimension_x, dimension_y, custom_plot = NULL)
}
\arguments{
\item{dimension_x}{A column symbol representing the X dimension. Not needed if using 
`custom_plot`.}

\item{dimension_y}{A column symbol representing the Y dimension. Not needed if using 
`custom_plot`.}

\item{custom_plot}{An optional ggplot object, allowing for custom interactive plots. Must contain 
a row index in the `.key` column.}
}
\value{
A vector with TRUE for elements inside gate points and FALSE for elements outside gate 
points. A record of the selected points is stored in `tidygate_env$select_data` and a 
record of the gates is stored in `tidygate_env$brush_data`.
}
\description{
Launch an interactive scatter plot, allowing manual selection of points based on X and Y
coordinates.
}
\examples{
\donttest{
library(dplyr)
library(ggplot2)

# Use with default plot
mtcars |>
  dplyr::mutate(selected = gate_interactive(dimension_x = mpg, dimension_y = wt)) |>
  print()

# Use with custom plot
scaled_plot <- 
  mtcars |> 
  dplyr::mutate(.key = dplyr::row_number()) |>
  ggplot2::ggplot(ggplot2::aes(x = mpg, y = wt, key = .key)) + 
  ggplot2::scale_y_log10() +
  ggplot2::geom_point() +
  ggplot2::theme_dark()
  
mtcars |>
  dplyr::mutate(selected = gate_interactive(dimension_x = mpg, dimension_y = wt, custom_plot = scaled_plot)) |>
  print()
}
}
