% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_interactive}
\alias{gate_interactive}
\title{Interactively gate data with a simple scatter plot}
\usage{
gate_interactive(x, y, colour = NULL, shape = NULL, alpha = 1, size = 2)
}
\arguments{
\item{x}{A vector representing the X dimension.}

\item{y}{A vector representing the Y dimension.}

\item{colour}{A vector representing the point colour. Or, a colour code string compatible
with ggplot2.}

\item{shape}{A vector representing the point shape, coercible to a factor. Or, a shape code 
numeric compatible with ggplot2 (0-127)}

\item{alpha}{A vector representing the point alpha, coercible to a factor. Or, an alpha numeric 
compatible with ggplot2 (0-1).}

\item{size}{A vector representing the point size, coercible to a factor. Or, a size numeric 
compatible with ggplot2 (0-20).}
}
\value{
A vector of strings, of the gates each X and Y coordinate pair is within. If gates are
drawn interactively, is stored as `tidygate_env$gates`.
}
\description{
Create an interactive scatter plot based on user-defined X and Y coordinates. Colour, shape, size 
and alpha can be defined as constant values, or can be controlled by values in a specified 
column.
}
\examples{
\dontrun{
mtcars |>
  mutate(selected = gate_interactive(x = mpg, y = wt, shape = am))
}
}
