% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_OLD.R
\name{gate_interactive}
\alias{gate_interactive}
\title{Get points within a user drawn gate}
\usage{
gate_interactive(
  .data,
  .element,
  .dim1,
  .dim2,
  .color = NULL,
  .shape = NULL,
  .size = NULL,
  opacity = 1,
  how_many_gates = 1,
  name = "gate",
  ...
)
}
\arguments{
\item{.data}{A tibble}

\item{.element}{A column symbol. The column that is used to calculate distance (i.e., normally genes)}

\item{.dim1}{A column symbol. The x dimension}

\item{.dim2}{A column symbol. The y dimension}

\item{.color}{A column symbol. Color of points}

\item{.shape}{A column symbol. Shape of points}

\item{.size}{A column symbol. Size of points}

\item{opacity}{A number between 0 and 1. The opacity level of the data points}

\item{how_many_gates}{An integer. The number of gates to label}

\item{name}{A character string. The name of the new column}

\item{...}{Further parameters passed to the function gatepoints::fhs}
}
\value{
A tibble with additional columns
}
\description{
Get points within a user drawn gate
}
\keyword{internal}
