% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_programmatic}
\alias{gate_programmatic}
\title{Get points within a user drawn gate}
\usage{
gate_programmatic(dimension_x, dimension_y, gate_list)
}
\arguments{
\item{dimension_x}{A column symbol representing the X dimension.}

\item{dimension_y}{A column symbol representing the Y dimension.}

\item{gate_list}{A list of of dataframes recording gate coordinates, as output by 
`tidygate_env$brush_data`.}
}
\value{
A vector of lists, recording the gates each X and Y coordinate pair is within.
}
\description{
Get points within a user drawn gate
}
