% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gate_programmatic}
\alias{gate_programmatic}
\title{Gate data with pre-recorded lasso selection coordinates}
\usage{
gate_programmatic(x_column, y_column, gates)
}
\arguments{
\item{x_column}{A column symbol representing the X dimension.}

\item{y_column}{A column symbol representing the Y dimension.}

\item{gates}{A `data.frame` recording the gate brush data, as output by 
`tidygate_env$brush_data`. The column `x` records X coordinates, the column `y` records Y 
coordinates and the column `.gate` records the gate.}
}
\value{
A vector of lists, of the gates each X and Y coordinate pair is within.
}
\description{
A helpful way to repeat previous interactive lasso selections to enable reproducibility. 
Programmatic gating is based on the package [gatepoints](https://github.com/wjawaid/gatepoints)
by Wajid Jawaid.
}
